delete bridge files in workspace
